requirements.txt
aiohttp
aio-pika
pillow